# Improvements to building and CI for mingw/msys