@Mark-Simulacrum 2024-02-18
cc @mati865 -- perhaps you can weigh in here? My knowledge of Windows isn't deep enough to say whether this moves us in the right direction or how closely affiliated the msys2 org/github action is with the upstream project (i.e. is trusting it equivalent to trusting msys2 artifacts?). We may still want to replicate the artifacts as I think we were doing before into our own cache regardless...