@mati865 2024-02-18
> how closely affiliated the msys2 org/github action is with the upstream project

`msys2/setup-msys2` is the official GH Action maintained by MSYS2 and the recommend way to install it.

> whether this moves us in the right direction

I'm currently ill and don't want to make any statements because of my impaired ability to understand the code.
Since the problem in #105696 is caused by MSYS2 git returning UNIX style paths (like `/c/rust/src/whatever`) I think it might be simpler to just call `cygpath` (with `-w` or `-m` depending or what bootstrap expects) on paths passed to obtained from git.
Conversion from UNIX to Windows (`cygpath -w`) would give `C:\rust\src\whatever` and conversion to mixed (`-cygpath -m`) would give give `C:/rust/src/whatever`.
Other alternative would be to use a library that would give bootstrap the proper path from the beginning like gix https://github.com/rust-lang/rust/pull/110369 or git2-rs.