@majaha 2024-02-18
I did think about Cygpath, but just getting git to give us the path of `../../../`s we need seemed like a simpler solution to me.