@mati865 2024-02-21
In overall looks good but I'm not sure whether Rust should install MSYS2 each time as it adds another possible point of failure to the CI. However if the caching works good this might not be a problem.