@Mark-Simulacrum 2024-02-24
I'm going to go ahead and approve. I'm not an expert on this area but my sense is that this should be easy to revert if needed, so that's my preferred approach here.

@bors r+