@ChrisDenton 2024-02-25
Just a note but it seems that the mingw builders are now the slowest by roughly 20 to 30 minutes.

|Job|Time|
|--|--|
|auto - i686-mingw|2h 22m|
|auto - x86_64-mingw|2h 20m|
|auto - dist-x86_64-linux-alt|1h 55m|

This is a fair increase from the roughly 1hr 40m it took before. I don't know if that was expected.