@mati865 2024-02-26
Agreed. Initially, I thought it might be a cache issue that would hopefully improve. But on second thought, I think it's just the slowness of Cygwin/MSYS2 Git.
I'll try to check the logs later today, but I think the revert is well justified.