@mati865 2024-02-27
Sorry, haven't managed to take a look yesterday.
Looking at `auto` branch build:
- before this PR: https://github.com/rust-lang-ci/rust/actions/runs/8040180131/job/21957991997
  - repo clone: 37s
  - submodules checkout: 1m 27s
  - tidy: `[TIMING] core::build_steps::test::Tidy Tidy -- 39.564`
  - bootstrap tests: `[TIMING] core::build_steps::test::Bootstrap Bootstrap -- 187.457`
- after the merge: https://github.com/rust-lang-ci/rust/actions/runs/8065622298/job/22031836054
  - repo clone: 1m 57s
  - submodules checkout: 3m 13s
  - MSYS2 install: 8m 34s, it's not MSYS2 that is slow but the script that removes preinstalled tools: https://github.com/rust-lang/rust/blob/ef324565d071c6d7e2477a195648549e33d6a465/src/ci/scripts/install-msys2.sh#L34-L39
  - tidy: `[TIMING] core::build_steps::test::Tidy Tidy -- 113.468`
  - bootstrap tests: `[TIMING] core::build_steps::test::Bootstrap Bootstrap -- 1939.995`

So indeed everything that uses Git became much slower and I don't see any way to regain the performance other than reverting CI changes part of this PR.