@majaha 2024-02-28
Unfortunately, MSYS's Git is just slow. I thought it would be neater if it were tested in CI so that we can confidently suggest using it in the build instructions. But if it really is unbearably slow, we could perhaps take another approach. Maybe we could tell people to steer clear of MSYS Git and just install Git-for-Windows instead, although that's a bit of a pain to have to install a separate thing.

I don't understand the difference in time for the "checkout the source code" step, there shouldn't be a difference. Although taking a closer look, maybe the "disable git crlf conversion" step isn't applying to the checkout action because the config is going in the wrong home directory, and the extra work of expanding the line endings accounts for the extra time taken? (We *could* run that script twice, once in MSYS Bash and once in Git Bash, to make sure it gets configured everywhere).

For `install-msys2.sh`, that script is also installing things from network, which could be slowing it down. But it's also quite possible that deleting all those files is also what's slowing it down. (Maybe we could just `mv` them off path instead?)

I think quite a lot of this would be sped up by just moving back to using Git-for-Windows for mingw builds, but if we do that we should probably list it as a stronger requirement in the build instructions.