@mati865 2024-03-01
> I don't understand the difference in time for the "checkout the source code" step, there shouldn't be a difference.

Sorry, I was sleepy and haven't realised MSYS2 is installed after the checkout. Longer checkout must have been a noise.

> For install-msys2.sh, that script is also installing things from network, which could be slowing it down. But it's also quite possible that deleting all those files is also what's slowing it down. (Maybe we could just mv them off path instead?)

If you enable timestamps you can see:
```
Tue, 27 Feb 2024 13:42:03 GMT installing pactoys...
Tue, 27 Feb 2024 13:49:37 GMT resolving dependencies...
Tue, 27 Feb 2024 13:49:38 GMT looking for conflicting packages...
Tue, 27 Feb 2024 13:49:38 GMT
Tue, 27 Feb 2024 13:49:38 GMT Packages (33) mingw-w64-x86_64-brotli-1.1.0-1 
```
Which means CI spent 7 minutes between lines https://github.com/rust-lang/rust/blob/ef324565d071c6d7e2477a195648549e33d6a465/src/ci/scripts/install-msys2.sh#L29 and https://github.com/rust-lang/rust/blob/ef324565d071c6d7e2477a195648549e33d6a465/src/ci/scripts/install-msys2.sh#L47
Pactoys install is fast because it's tiny so calling `rm` is the only thing left.

> I think quite a lot of this would be sped up by just moving back to using Git-for-Windows for mingw builds, but if we do that we should probably list it as a stronger requirement in the build instructions.

I'll revive https://github.com/rust-lang/rust/pull/110369 over the weekend and see if build time of newer version is acceptable. It should be even faster than Git for Windows.