@mati865 2024-03-04
> I'll revive https://github.com/rust-lang/rust/pull/110369 over the weekend and see if build time of newer version is acceptable. It should be even faster than Git for Windows.

With the latest version the time goes from 10s to 13s on my PC so about 30% increase, definitely an improvement but still not ideal.

@majaha are you going to revert CI changes?