@majaha 2024-03-05
> With the latest version the time goes from 10s to 13s on my PC so about 30% increase, definitely an improvement but still not ideal.
Sorry, could you explain what those numbers were measurements of?

I'd rather not revert all the CI changes as I think parts of it are more sensible than what was there before. But I'm willing to try and fix up the main issues, if you wouldn't mind pointing me to what you think they are.

The biggest issue is this: `bootstrap tests: [TIMING] core::build_steps::test::Bootstrap Bootstrap -- 1939.995`
Running the tests for bootstrap takes over 30 minutes by itself. I've noticed that bootstrap became excruciatingly slow with MSYS's git in my own testing. I'm sure this will be fixed by reverting back to Git-for-Windows, so I think we should start there. I'll try to make a pull request for that tomorrow evening.